+++ CHANGELOG.md
@@ -10,14 +10,14 @@ Node.js 12 is currently not supported by this plugin due to compatibility issues
 
 ### Notable Changes
 
-* **Plugins:** Allow plugins to specify only a single config block should exist in the GUI config editor ([#290](https://github.com/oznu/homebridge-config-ui-x/issues/290))
+* **Plugins:** Allow plugins to specify only a single config block should exist in the Plugin Setting GUI ([#290](https://github.com/oznu/homebridge-config-ui-x/issues/290))
 * **Plugins:** The plugin will no longer log with timestamps when homebridge is started with `-T` ([#288](https://github.com/oznu/homebridge-config-ui-x/issues/288))
 * **Plugins:** The plugin logs are now prefixed with the name defined in the `config.json` ([#288](https://github.com/oznu/homebridge-config-ui-x/issues/288))
 * **Auth:** Display a warning if the time on the server varies to greatly from time on the client which could cause mysterious login issues
 
 ### Bug Fixes
 
-* **Plugins:** Fixed a bug preventing users adding additional elements to an array with the GUI config editor ([#289](https://github.com/oznu/homebridge-config-ui-x/issues/289))
+* **Plugins:** Fixed a bug preventing users adding additional elements to an array with the Plugin Setting GUI ([#289](https://github.com/oznu/homebridge-config-ui-x/issues/289))
 * **Plugins:** Fixed an issue preventing the plugins tab from loading when the npm registry was slow to respond ([#284](https://github.com/oznu/homebridge-config-ui-x/issues/284))
 * **Plugins:** Fixed an issue that prevented plugins from displaying for some users ([#284](https://github.com/oznu/homebridge-config-ui-x/issues/284))
 * **Plugins:** Searching for the exact name of a plugin will return a result even if the [npm registry search is not working](https://status.npmjs.org/incidents/qg46dsfk1vt2)